backend/requirements.txt
# --- Core Backend ---
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2
pydantic==2.12.5
python-dotenv==1.2.1

# --- Database ---
motor==3.3.1
pymongo==4.5.0

# --- HTTP / Gemini REST ---
requests==2.32.5

# --- Utilities ---
python-multipart==0.0.21
email-validator==2.3.0
bcrypt==4.1.3
passlib==1.7.4

# --- Async / Networking ---
anyio==4.12.0
httpx==0.28.1
aiohttp==3.13.3

# --- Dev / Logging (optional but safe) ---
rich==14.2.0
